重入id扩展为int

+++ java/BTree-Core/src/main/java/cn/wjybxx/btree/Task.java
@@ -128,7 +128,7 @@ public abstract class Task<T> implements ICancelTokenListener {
     /** 任务运行时的控制信息(bits) -- 每次运行时会重置为0 */
     private transient int ctl;
     /** 重入Id，只增不减 -- 用于事件驱动下检测冲突（递归）；reset时不重置，甚至也增加 */
-    private transient short reentryId;
+    private transient int reentryId;
 
     /** 节点的名字 */
     protected String name;
@@ -789,7 +789,7 @@ public abstract class Task<T> implements ICancelTokenListener {
         initMask |= (prevStatus << OFFSET_PREV_STATUS);
         ctl = initMask;
         status = TaskStatus.RUNNING; // 先更新为running状态，以避免执行过程中外部查询task的状态时仍处于上一次的结束status
-        final short reentryId = ++this.reentryId;  // 和上次执行的exit分开
+        final int reentryId = ++this.reentryId;  // 和上次执行的exit分开
         // beforeEnter
         if ((initMask & TaskOverrides.MASK_BEFORE_ENTER) != 0) {
             beforeEnter(); // 这里用户可能修改控制流标记
@@ -851,7 +851,7 @@ public abstract class Task<T> implements ICancelTokenListener {
             setCompleted(TaskStatus.CANCELLED, false);
             return;
         }
-        final short reentryId = this.reentryId;
+        final int reentryId = this.reentryId;
         execute();
         if (reentryId != this.reentryId) {
             return;
@@ -878,8 +878,8 @@ public abstract class Task<T> implements ICancelTokenListener {
             return;
         }
         // 如果source收到取消信号，则被内联的节点的子节点也一定收到取消信号
-        final short sourceReentryId = source.reentryId;
-        final short reentryId = this.reentryId;
+        final int sourceReentryId = source.reentryId;
+        final int reentryId = this.reentryId;
         // 内联template_execute逻辑
         outer:
         {
